seperate ReportProblemModal to its own file
